cinegram/requirements.txt
python-telegram-bot[job-queue]
requests
Pillow
python-dotenv

guessit
